Linked authentication with landing page and feed
